+++ uui/src/lib/Select/Select.svelte
@@ -1,0 +1,78 @@
+<script lang="ts">
+  import TextField from '../TextField/TextField.svelte';
+  import Box from '../Box/box.svelte';
+  import Menu from '../Menu/Menu.svelte';
+  import type { SelectProps, SelectSlots } from './Select.types';
+
+  // Props
+  type $$Slots = SelectSlots;
+  type $$Props = SelectProps;
+  export let multiple = false;
+  export let chip = false;
+  export let labelId = '';
+  export let open = false;
+  export let value = '';
+
+  // Local Vars
+  let tfContainerRef: HTMLElement;
+  let inputRef: HTMLInputElement;
+  let helperTextOffset = -18;
+
+  // Functions
+
+  // Passed as Slot Prop for trailing icon to use on:click
+  function setOpenANDFocus() {
+    open = !open;
+
+    if (open && document.activeElement != inputRef) {
+      inputRef.focus();
+    }
+  }
+
+  // Passed as Slot Prop in default slot of Menu,
+  // which means "menu" items can use this on:click
+  function handleClicked(e: MouseEvent) {
+    value = (e.target as HTMLElement).textContent?.trim() || '';
+    open = false;
+  }
+</script>
+
+<Box class="select {$$restProps.class || ''}">
+  <TextField
+    bind:ref={tfContainerRef}
+    bind:inputRef
+    on:focus={() => {
+      open = true;
+    }}
+    {value}
+    readonly
+    variant={$$restProps.variant || 'outlined'}
+    label={$$restProps.label}
+    helperText={$$restProps.helperText || ''}
+  >
+    <slot name="trailing" slot="trailing" setOpen={setOpenANDFocus} />
+  </TextField>
+
+  <Menu
+    bind:open
+    bind:anchorRef={tfContainerRef}
+    anchorOrigin={$$restProps.anchorOrigin || 'bottom-start'}
+    offsetNum={helperTextOffset}
+  >
+    <slot {handleClicked} />
+  </Menu>
+</Box>
+
+<style lang="scss" global>
+  .select {
+    width: fit-content;
+  }
+  .select .text-field .mdc-text-field__input {
+    cursor: default;
+    caret-color: transparent;
+  }
+
+  .select .menu {
+    border-radius: 0px 0px 4px;
+  }
+</style>