+++ uui/src/lib/TextField/TextField.types.d.ts
@@ -3,18 +3,18 @@ import type { ThemeColors } from '../shared/theme/default-colors/colors.types';
 
 export type TextFieldVariant = 'outlined' | 'filled';
 
-export interface TextFieldProps extends svelte.JSX.HTMLAttributes<HTMLInputElement> {
-  variant: TextFieldVariant;
+export interface TextFieldProps
+  extends BoxProps<HTMLInputElement>,
+    svelte.JSX.HTMLAttributes<HTMLInputElement> {
+  variant?: TextFieldVariant;
   label: string;
   value?: string;
   helperText?: string;
   color?: ThemeColors;
   error?: boolean;
   characterCounter?: boolean;
+  inputRef?: HTMLInputElement;
+  clickaway?: boolean;
 }
 
-export type TextFieldIconVariant = 'leading' | 'trailing';
-
-export interface TextFieldIconProps<T extends EventTarget> extends BoxProps<T> {
-  variant: TextFieldIconVariant;
-}
+export type TextFieldIconProps<T extends EventTarget> = BoxProps<T>;